:recycle:: Clarify service responsibilities for user management
